+++ src/rich_toolkit/app_style.py
@@ -1,9 +1,8 @@
 from abc import ABC, abstractmethod
 from typing import Any, Generator, List, Union
 
-from rich._loop import loop_last
+from rich._loop import loop_first_last
 from rich.color import Color
-from rich.color_triplet import ColorTriplet
 from rich.console import (
     Console,
     ConsoleOptions,
@@ -15,7 +14,7 @@ from rich.segment import Segment
 from rich.style import Style
 from rich.text import Text
 
-RenderDecorationResult = Generator[Segment, None, None]
+from .utils.colors import lighten
 
 
 class AppStyle(ABC):
@@ -42,7 +41,7 @@ class AppStyle(ABC):
 
         self._animation_counter = 0
 
-    def render_empty_line(self) -> Text:
+    def empty_line(self) -> Text:
         return Text(" ")
 
     def with_decoration(
@@ -66,20 +65,8 @@ class AppStyle(ABC):
     @abstractmethod
     def decorate(
         self, lines: List[List[Segment]], animated: bool = False, **kwargs: Any
-    ) -> RenderDecorationResult:
+    ) -> Generator[Segment, None, None]:
         raise NotImplementedError()
-
-
-def lighten(color: Color, amount: float) -> Color:
-    assert color.triplet
-
-    r, g, b = color.triplet
-
-    r = int(r + (255 - r) * amount)
-    g = int(g + (255 - g) * amount)
-    b = int(b + (255 - b) * amount)
-
-    return Color.from_triplet(ColorTriplet(r, g, b))
 
 
 class TaggedAppStyle(AppStyle):
@@ -88,7 +75,9 @@ class TaggedAppStyle(AppStyle):
 
         super().__init__(*args, **kwargs)
 
-    def _render_tag(self, text: str, background_color: Color) -> RenderDecorationResult:
+    def _render_tag(
+        self, text: str, background_color: Color
+    ) -> Generator[Segment, None, None]:
         style = Style.from_color(Color.parse("#ffffff"), bgcolor=background_color)
 
         if text:
@@ -103,7 +92,7 @@ class TaggedAppStyle(AppStyle):
 
     def decorate(
         self, lines: List[List[Segment]], animated: bool = False, **kwargs: Any
-    ) -> RenderDecorationResult:
+    ) -> Generator[Segment, None, None]:
         if animated:
             yield from self.decorate_with_animation(lines)
 
@@ -113,15 +102,17 @@ class TaggedAppStyle(AppStyle):
 
         color = self.highlight_color if kwargs.get("title", False) else self.base_color
 
-        for index, line in enumerate(lines):
-            text = tag if index == 0 else ""
+        for first, last, line in loop_first_last(lines):
+            text = tag if first else ""
             yield from self._render_tag(text, background_color=color)
             yield from line
-            yield Segment.line()
+
+            if not last:
+                yield Segment.line()
 
     def decorate_with_animation(
         self, lines: List[List[Segment]]
-    ) -> RenderDecorationResult:
+    ) -> Generator[Segment, None, None]:
         block = "█"
 
         block_length = 5
@@ -133,8 +124,8 @@ class TaggedAppStyle(AppStyle):
 
         self._animation_counter += 1
 
-        for index, line in enumerate(lines):
-            if index == 0:
+        for first, last, line in loop_first_last(lines):
+            if first:
                 yield Segment(" " * left_padding)
 
                 for j in range(block_length):
@@ -152,7 +143,7 @@ class TaggedAppStyle(AppStyle):
 class FancyAppStyle(AppStyle):
     def decorate(
         self, lines: List[List[Segment]], animated: bool = False, **kwargs: Any
-    ) -> RenderDecorationResult:
+    ) -> Generator[Segment, None, None]:
         if animated:
             colors = [lighten(self.base_color, 0.1 * i) for i in range(0, 5)]
 
@@ -160,8 +151,8 @@ class FancyAppStyle(AppStyle):
 
             color_index = self._animation_counter % len(colors)
 
-            for index, (last, line) in enumerate(loop_last(lines)):
-                if index == 0:
+            for first, last, line in loop_first_last(lines):
+                if first:
                     yield Segment("◆ ", style=Style(color=colors[color_index]))
                 else:
                     yield Segment("  ")
@@ -170,8 +161,8 @@ class FancyAppStyle(AppStyle):
 
             return
 
-        for index, (last, line) in enumerate(loop_last(lines)):
-            if index == 0:
+        for first, last, line in loop_first_last(lines):
+            if first:
                 decoration = "┌ " if kwargs.get("title", False) else "◆ "
             elif last:
                 decoration = "└ "
@@ -180,7 +171,9 @@ class FancyAppStyle(AppStyle):
 
             yield Segment(decoration)
             yield from line
-            yield Segment.line()
 
-    def render_empty_line(self) -> Text:
+            if not last:
+                yield Segment.line()
+
+    def empty_line(self) -> Text:
         return Text("│")
